paramsbuilder uses provider package to figure out HTTP Client
